+++ changelog.md
@@ -7,11 +7,13 @@ Make sure to report issues you would encounter, thank you :)
 
 New effects:
 - New SAO: Death Knight's Sudden Doom (Cataclysm)
+- New SAO: Death Knight's Will of the Necropolis (Cataclysm)
 - New SAO: Hunter's Lock and Load (Season of Discovery)
 - New SAO: Warrior's Sword and Board (Season of Discovery)
 - New GAB: Death Knight's Death Coil, during Sudden Doom (Cataclysm)
 - New GAB: Death Knight's Icy Touch, during Rime (Cataclysm)
 - New GAB: Death Knight's Obliterate, during Killing Machine (Cataclysm)
+- New GAB: Death Knight's Rune Tap, during Will of the Necropolis (Cataclysm)
 - New GAB: Paladin's Holy Shock, as combat-only counter
 - New GAB: Warrior's Overpower, during Taste for Blood* (SoD) 
 - New GAB: Warrior's Slam, during Sword and Board (SoD)